refactor: move table files to module folders
