+++ src/modules/clubs/table/row.tsx
@@ -1,10 +1,10 @@
 import { Delete as DeleteIcon, Edit as EditIcon } from '@mui/icons-material'
 import { useRouter } from 'next/router'
 
-import { StyledTableCell } from '@/components/tables/common/cell'
-import { TableMenu } from '@/components/tables/common/menu'
-import { TableMenuItem } from '@/components/tables/common/menu-item'
-import { StyledTableRow } from '@/components/tables/common/row'
+import { StyledTableCell } from '@/components/tables/cell'
+import { TableMenu } from '@/components/tables/menu'
+import { TableMenuItem } from '@/components/tables/menu-item'
+import { StyledTableRow } from '@/components/tables/row'
 import { ClubDto } from '@/modules/clubs/types'
 import { getFlagEmoji } from '@/utils/get-flag-emoji'
 import { useTableMenu } from '@/utils/hooks/use-table-menu'
