+++ src/modules/players/table/row.tsx
@@ -11,14 +11,14 @@ import Link from 'next/link'
 import { useRouter } from 'next/router'
 import { useTranslation } from 'next-i18next'
 
-import { PlayerDto } from '@/modules/players/types'
+import { StyledTableCell } from '@/components/tables/cell'
+import { TableMenu } from '@/components/tables/menu'
+import { TableMenuItem } from '@/components/tables/menu-item'
+import { StyledTableRow } from '@/components/tables/row'
 import { getFlagEmoji } from '@/utils/get-flag-emoji'
 import { useTableMenu } from '@/utils/hooks/use-table-menu'
 
-import { StyledTableCell } from '../common/cell'
-import { TableMenu } from '../common/menu'
-import { TableMenuItem } from '../common/menu-item'
-import { StyledTableRow } from '../common/row'
+import { PlayerDto } from '../types'
 
 interface IPlayersTableRowProps {
   data: PlayerDto
