+++ src/pages/clubs/index.tsx
@@ -7,11 +7,11 @@ import { Fab } from '@/components/fab/fab'
 import { Loader } from '@/components/loader/loader'
 import { ConfirmationModal } from '@/components/modals/confirmation-modal'
 import { PageHeading } from '@/components/page-heading/page-heading'
-import { ClubsTable } from '@/components/tables/clubs'
-import { ClubsTableRow } from '@/components/tables/rows/clubs-row'
 import { withSessionSsr } from '@/modules/auth/session'
 import { ClubsFilterForm } from '@/modules/clubs/forms/filter'
 import { useClubs, useDeleteClub } from '@/modules/clubs/hooks'
+import { ClubsTableRow } from '@/modules/clubs/table/row'
+import { ClubsTable } from '@/modules/clubs/table/table'
 import { ClubsFiltersDto, ClubsSortBy } from '@/modules/clubs/types'
 import { useCountriesList } from '@/modules/countries/hooks'
 import { useRegionsList } from '@/modules/regions/hooks'
