+++ src/pages/players/[slug].tsx
@@ -4,13 +4,13 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
 import { ErrorContent } from '@/components/error/error-content'
 import { PageHeading } from '@/components/page-heading/page-heading'
-import { TeamAffiliationsTableRow } from '@/components/tables/rows/team-affiliations-row'
-import { TeamAffiliationsTable } from '@/components/tables/team-affiliations'
 import { withSessionSsr } from '@/modules/auth/session'
 import { PlayerDetialsCard } from '@/modules/players/details-card'
 import { PlayerDto } from '@/modules/players/types'
-import { useTeamAffiliations } from '@/modules/team-affiliations.ts/hooks'
-import { TeamAffiliationsSortBy } from '@/modules/team-affiliations.ts/types'
+import { useTeamAffiliations } from '@/modules/team-affiliations/hooks'
+import { TeamAffiliationsTableRow } from '@/modules/team-affiliations/table/row'
+import { TeamAffiliationsTable } from '@/modules/team-affiliations/table/team'
+import { TeamAffiliationsSortBy } from '@/modules/team-affiliations/types'
 import { getPlayerBySlug } from '@/services/api/methods/players'
 import { ApiError } from '@/services/api/types'
 import { useTable } from '@/utils/hooks/use-table'
