+++ src/pages/teams/[slug].tsx
@@ -4,10 +4,10 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
 import { ErrorContent } from '@/components/error/error-content'
 import { PageHeading } from '@/components/page-heading/page-heading'
-import { CompetitionParticipationsTable } from '@/components/tables/competition-participations'
-import { CompetitionParticipationsTableRow } from '@/components/tables/rows/competition-participations-row'
 import { withSessionSsr } from '@/modules/auth/session'
 import { useCompetitionParticipations } from '@/modules/competition-participations/hooks'
+import { CompetitionParticipationsTableRow } from '@/modules/competition-participations/table/row'
+import { CompetitionParticipationsTable } from '@/modules/competition-participations/table/table'
 import { CompetitionParticipationsSortBy } from '@/modules/competition-participations/types'
 import { TeamDetailsCard } from '@/modules/teams/details-card'
 import { TeamDto } from '@/modules/teams/types'
