+++ src/services/api/methods/team-affiliations.ts
@@ -1,7 +1,7 @@
 import {
   FindAllTeamAffiliationsParams,
   TeamAffiliationDto,
-} from '@/modules/team-affiliations.ts/types'
+} from '@/modules/team-affiliations/types'
 import { getPaginatedData } from '@/services/api/methods/helpers'
 import { TModuleName } from '@/services/api/modules'
